fixed proptype bug relates #116
