+++ client/src/components/application/NonlinearStepper/style.js
@@ -1,0 +1,14 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme) => ({
+  gender: {
+    display: 'flex',
+    justifyContent: 'flex-end',
+    margin: theme.spacing(3),
+  },
+  nextButton: {
+    marginLeft: theme.spacing(4),
+  },
+}));
+
+export default useStyles;
